requirements.txt
asgiref==3.8.1
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
pillow==11.0.0
psycopg==3.2.3
psycopg-binary==3.2.3
sqlparse==0.5.3
